refactor: storage card
