+++ components/ui/variants.tsx
@@ -117,24 +117,3 @@ export const BadgeVariants = cva(
 )
 
 export type BadgeProps = VariantProps<typeof BadgeVariants>
-
-export const cardVariants = cva(
-  "rounded-xl bg-white text-black transition-shadow",
-  {
-    variants: {
-      variant: {
-        default:
-          "border border-neutral-300 bg-gradient-to-br from-white to-neutral-50/30 rounded-2xl",
-        people: "border-none",
-        content: "w-full border-none shadow-none rounded-none px-0",
-      },
-      size: {
-        fit: "w-fit",
-        contained: "w-full max-w-sm mx-auto",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-    },
-  }
-)
